Refactor API and Svelte app into focused modules
