+++ backend/src/gpx_helper/api/config.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+API_VERSION = "v1"
+DEFAULT_ALLOWED_ORIGINS = (
+    "http://localhost:5173",
+    "http://localhost:4173",
+)
+MAX_UPLOAD_READ_BYTES = 1024 * 1024
