+++ frontend/src/components/AboutPage.svelte
@@ -1,0 +1,43 @@
+<section class="about-grid">
+  <article class="tool-card about-card">
+    <header class="section-header">
+      <h2>What this app does</h2>
+    </header>
+    <p class="muted-text">
+      GPX Helper trims GPX tracks and renders both route videos and telemetry overlays from the same browser UI.
+    </p>
+  </article>
+
+  <article class="tool-card about-card">
+    <header class="section-header">
+      <h2>Trim GPX</h2>
+    </header>
+    <ul class="about-list">
+      <li><strong>Trim by time:</strong> export one GPX file from a start and end time.</li>
+      <li><strong>Split by videos:</strong> create one GPX segment per video and download them as a ZIP.</li>
+      <li><strong>Optional video metadata:</strong> can fill clip times automatically when available.</li>
+    </ul>
+  </article>
+
+  <article class="tool-card about-card">
+    <header class="section-header">
+      <h2>Route animation</h2>
+    </header>
+    <ul class="about-list">
+      <li>Render the GPX route as an MP4.</li>
+      <li>Adjust duration, frame rate, size, tiles, and route styling.</li>
+      <li>Use it when you need a clean route preview or shareable map video.</li>
+    </ul>
+  </article>
+
+  <article class="tool-card about-card">
+    <header class="section-header">
+      <h2>Telemetry video</h2>
+    </header>
+    <ul class="about-list">
+      <li>Render telemetry-only MP4 overlays directly from a GPX file.</li>
+      <li>Choose elevation value, speed, heart rate, or an elevation graph.</li>
+      <li>Set duration, frame rate, size, and background color for compositing over footage.</li>
+    </ul>
+  </article>
+</section>
